Logic and hardware version of shiftregistersign
These files have almost no relation to the coded part of this project. For now this is what is used to make the sign work, i.e schematic and etc. The coded part will soon be edited and improved upon. But for now I have used IC's, timers and etc. to create patterns on the led array